requirements.txt
openai>=1.0.0
openai-agents>=0.0.15
pydantic>=2.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sendgrid>=6.0.0
gradio>=4.0.0
